Move DbGroupLightCrossref to local (database) package.
